Magento Fork initial commit

+++ Controller/Adminhtml/Wysiwyg/Images/DeleteFolder.php
@@ -4,18 +4,14 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-
-declare(strict_types=1);
-
 namespace Magento\Cms\Controller\Adminhtml\Wysiwyg\Images;
 
-use Magento\Framework\App\Action\HttpPostActionInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
 
 /**
  * Delete image folder.
  */
-class DeleteFolder extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images implements HttpPostActionInterface
+class DeleteFolder extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images
 {
     /**
      * @var \Magento\Framework\Controller\Result\JsonFactory
@@ -63,8 +59,13 @@ class DeleteFolder extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images impl
     {
         try {
             $path = $this->getStorage()->getCmsWysiwygImages()->getCurrentPath();
+            if (!$this->directoryResolver->validatePath($path, DirectoryList::MEDIA)) {
+                throw new \Magento\Framework\Exception\LocalizedException(
+                    __('Directory %1 is not under storage root path.', $path)
+                );
+            }
             $this->getStorage()->deleteDirectory($path);
-
+            
             return $this->resultRawFactory->create();
         } catch (\Exception $e) {
             $result = ['error' => true, 'message' => $e->getMessage()];
